Reducing the number of calls to MEZO precompile
In Mezo chain there is a limit of 10 calls per transaction to a given
precompile, which in our case is MEZO token. Here we are aiming at
reducing these calls by hardcoding MEZO decimals to 18 because we know
for a fact that MEZO has 18 decimals. And also we know that MEZO is not
a fee-on-transfer token, hence we can rely on amount that was
transferred from a sender.